+++ src/main/kotlin/com/example/history/application/inputs/CreateHistoryInput.kt
