+++ src/main/kotlin/com/example/history/application/services/GetAllHistoriesService.kt
