+++ src/main/kotlin/com/example/history/presentation/interfaces/api/DeleteHistoryController.kt
