+++ src/main/resources/application.properties
@@ -1,0 +1,15 @@
+server.port=8080
+
+spring.datasource.driverClassName=org.mariadb.jdbc.Driver
+spring.datasource.url=jdbc:mariadb://127.0.0.1:3306/history
+spring.datasource.username=root
+spring.datasource.password=secret
+
+spring.jpa.hibernate.ddl-auto=create-drop
+
+spring.flyway.url=jdbc:mariadb://127.0.0.1:3306/history
+spring.flyway.create-schemas=false
+spring.flyway.schemas=history
+spring.flyway.default-schema=history
+spring.flyway.user=root
+spring.flyway.password=secret
